pyproject.toml
[project]
name = "memadrift"
version = "0.2.0"
description = "Drift detection and remediation for Claude memory files"
readme = "README.md"
authors = [
    { name = "Sam Scarrow", email = "sam@scarrow.net" }
]
requires-python = ">=3.13"
dependencies = [
    "click>=8.1",
    "pyyaml>=6.0",
]

[project.scripts]
memadrift = "memadrift.cli:cli"

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["uv_build>=0.9.30,<0.10.0"]
build-backend = "uv_build"
